Refactor and enhance unit tests across multiple modules
- Updated ScrollState.cpp to use std::uint16_t for clamped speed.
- Changed drawXbm565 function signature in dino_game.cpp and main.cpp to accept const uint8_t* instead of const char*.
- Migrated color_parser_test.cpp from gtest to Unity framework, updating test cases accordingly.
- Added main function to various test files to facilitate Unity test execution.
- Improved bitmap tests in bitmapTest.cpp for clarity and correctness.
- Refactored test_ble.cpp, test_easing_functions.cpp, test_main.cpp, test_microphone.cpp, and test_scroll_state.cpp to include main functions and proper setup/teardown.
- Introduced new tests for getCurrentThreshold() and onSoftMillisTimer() in test_threshold_timer.cpp.
- Created a placeholder test for microphone functionality in test_mic.cpp.

+++ src/core/ScrollState.cpp
@@ -27,7 +27,7 @@ std::uint16_t clampSpeed(std::uint16_t speed)
 
 std::uint16_t computeInterval(std::uint16_t speed)
 {
-  const std::uint8_t clampedSpeed = clampSpeed(speed);
+  const std::uint16_t clampedSpeed = clampSpeed(speed);
   const std::uint16_t range = kSlowestInterval - kFastestInterval;
   const std::uint16_t steps = static_cast<std::uint16_t>(kMaxSpeed - kMinSpeed);
   const std::uint16_t offset = static_cast<std::uint16_t>(
